appliance/postgresql: Move new packages into place, remove old package
