+++ appliance/postgresql/.gitignore
@@ -1,1 +1,1 @@
-postgresql
+cmd/simcli/simcli
