--- appliance/postgresql/pg_hba.conf
@@ -1,4 +1,0 @@
-# TYPE  DATABASE        USER            ADDRESS                 METHOD
-local   all             all                                     peer
-host    all             all             0.0.0.0/0               md5
-host    replication     flynn           0.0.0.0/0               md5
