+++ appliance/postgresql/postgres.go
@@ -17,8 +17,8 @@ import (
 
 	"github.com/flynn/flynn/Godeps/_workspace/src/github.com/jackc/pgx"
 	"github.com/flynn/flynn/Godeps/_workspace/src/gopkg.in/inconshreveable/log15.v2"
-	"github.com/flynn/flynn/appliance/postgresql2/state"
-	"github.com/flynn/flynn/appliance/postgresql2/xlog"
+	"github.com/flynn/flynn/appliance/postgresql/state"
+	"github.com/flynn/flynn/appliance/postgresql/xlog"
 	"github.com/flynn/flynn/discoverd/client"
 	"github.com/flynn/flynn/pkg/shutdown"
 )
