+++ appliance/postgresql/simulator/simulator.go
@@ -28,8 +28,8 @@ import (
 	"time"
 
 	"github.com/flynn/flynn/Godeps/_workspace/src/gopkg.in/inconshreveable/log15.v2"
-	"github.com/flynn/flynn/appliance/postgresql2/state"
-	"github.com/flynn/flynn/appliance/postgresql2/xlog"
+	"github.com/flynn/flynn/appliance/postgresql/state"
+	"github.com/flynn/flynn/appliance/postgresql/xlog"
 	"github.com/flynn/flynn/discoverd/client"
 )
