--- appliance/postgresql2/.gitignore
@@ -1,1 +1,0 @@
-cmd/simcli/simcli
